Convert to coroutines. Start to map markers
Also cleans up code, optimize imports,

+++ app/src/main/java/xyz/wim/dentrealitytask/data/Country.kt
@@ -4,7 +4,7 @@ import com.squareup.moshi.Json
 
 data class Country(
     val timezones: List<String>,
-    val latlng: List<Float>,
+    val latlng: List<Double>,
     val name: String,
     @Json(name = "country_code") val countryCode: String,
     val capital: String
